requirements_kl.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.103.0
filelock==3.14.0
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.42.0
loguru==0.7.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.33.0
openai-whisper==20231117
PyAudio==0.2.14
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
starlette==0.29.0
sympy==1.12.1
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.23.1
deepgram-sdk==0.11.2
